Rewrite isNumber, re-build dist
