--- utils/is-number.js
@@ -1,14 +1,0 @@
-/*  
-Copyright (c) 2014-present, Jon Schlinkert, 
-<https://github.com/jonschlinkert/is-number> 
-*/
-
-export function isNumber(num) {
-  if (typeof num === 'number') {
-    return num - num === 0;
-  }
-  if (typeof num === 'string' && num.trim() !== '') {
-    return Number.isFinite ? Number.isFinite(+num) : isFinite(+num);
-  }
-  return false;
-}
