+++ webpack.config.js
@@ -4,6 +4,7 @@ module.exports = {
   mode: 'production',
   entry: './src/index.js',
   output: {
+    globalObject: 'this',
     path: path.resolve(__dirname, 'dist'),
     filename: 'linear-array.js',
     library: 'linearArray',
